Git Commands Every Developer Should Know
Git is an important part of daily programming (especially if you're working with a team) and is widely used in the software industry.
Since there are many various commands you can use, mastering Git takes time. But some commands are used more frequently (some daily). So in this post, I will share and explain the 10 most used Git commands that every developer should know.
Note: To understand this article, you need to know the basics of Git.